Create separate file for StreamingComponentExecutor class. Add RetryStrategy.
